Combine AuthWrapper and auth provider
